Create Python password hasher
This commit introduces the main file, the actual hasher. Fairly simple
as to how it works, but essentially the script uses crypt.crypt() to
create the password hash.

If you're using Python 3 (which you should be, cause Buhman says so),
then you can use crypt.mksalt() to make the salt. This actually has that
commented out, because the systems I'm using uses Python 2.7 by default.
However, mksalt is probably more betterer than os.urandom().

That being said, there is a cryptographically secure alternative for
Python 2.7, which we're using. The drawback is you have to import two
more modules with this alternative (os and binascii). We use
os.urandom() for Python 2.7ers.

Signed-off-by: Dave Russell Jr <dave@createazure.com>